Fixed many speed issues.
Modifed the addToHierarchy method: No longer uses an evolution model but rather
just tries to add the relationship. If it fails, the next relationship is added.
This makes adding relationships extremely fast.

Modified the term extractor methods: The speed went slower and slower since all
the stems and synonyms in the database needed to be loaded for each extraction.
Changed it so that they are loaded at the very beginning. This still needs to be
modified so that the whole database does not need to be loaded.

Modified the main method so that a directory may be passed as an argument. If this
is true, all the files in the directory are passed through the extraction algorithms.